Safari-609.1.11.1/LayoutTests/platform/wk2/http/tests/contentextensions/block-everything-unless-domain-expected.txt

CONSOLE MESSAGE: line 5: helper loaded
CONSOLE MESSAGE: Content blocker prevented frame displaying http://localhost:8000/contentextensions/resources/should-not-load.html from loading a resource from http://localhost:8000/contentextensions/resources/should-not-load.html
CONSOLE MESSAGE: line 13: PASS - blocked the load
